+++ src/modules/cars/repositories/implementations/CategoryRepository.ts
@@ -1,5 +1,5 @@
-import { Category } from "../model/category";
-import { ICategoriesRepository,ICreateCategoryDTO  } from "./ICategoriesRepository";
+import { Category } from "../../model/category";
+import { ICategoriesRepository,ICreateCategoryDTO  } from "../ICategoriesRepository";
 
 //DTO
