+++ src/modules/cars/repositories/implementations/SpecificationsRepository.ts
@@ -1,5 +1,5 @@
-import { ICreateSpecificationDTO, ISpecificationRepository } from "./ISpecificationRepository";
-import {Specification} from '../model/Specification'
+import { ICreateSpecificationDTO, ISpecificationRepository } from "../ISpecificationRepository";
+import {Specification} from '../../model/Specification'
 
 class SpecificationsRepository implements ISpecificationRepository {
   private specifications: Specification[];
