+++ src/routes/specifications.routes.ts
@@ -1,7 +1,7 @@
 import {Router} from 'express';
 
 import {CreateSpecificationService} from '../modules/cars/services/CreateSpecificationService'
-import {SpecificationsRepository} from '../modules/cars/repositories/SpecificationsRepository';
+import {SpecificationsRepository} from '../modules/cars/repositories/implementations/SpecificationsRepository';
 
 const specificationsRoutes = Router();
 const specificationRepository = new SpecificationsRepository();
